ai-training/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
tensorflow>=2.15.0,<2.20.0
opencv-python>=4.8.0
numpy>=1.24.0,<1.26.0
pillow>=9.5.0,<10.0.0
supabase>=2.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0